Add a fuse filesystem so that all files, regardless of assetstore, can be accessed.
This requires girder PR #1855.

This is a WIP.  Unmounting is not as clean as it could be (and, if Girder is killed, the mount causes Girder to linger until the mount point is accessed in some manner).  It is using a config file value, where perhaps it should use a system setting.  There are no tests.  There is a defect where if an item has two files with the same name, access to them is ambiguous.  It would be nice to use a single set of functions for this and some of the sftp module (such as determining access information).