Add an upload-annotations menu item.
This example shows how to add to the checked menu widget and upload
annotations to a single picked item.  It could use better feedback,
especially on large annotation sets.

This also makes a slight change to the annotation/item/:id endpoint so
that a single annotation could be passed rather than requiring only
lists of annotations.